Extract `SettingsUpgrader` implementations to separate files
